requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
openai==1.54.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
httpx==0.27.2
requests==2.32.3
bcrypt==4.2.0
email-validator==2.3.0
bleach==6.1.0
Pillow==10.1.0
